fix argument escaping for wget-Headers
Since executeProcess seems to be similar to IO.popen with regards to
argument/parameter handling we can't actually escape arguments unless
the command requires us to do so. wget sends the headers without change
to the server which will result in an error since accept headers are not
to be grouped by quotation marks.