check only if unexpanded IRIs are predefined
XML creates expanded IRIs without a prefix.
"function Expand" checked for a predefined local part
(and a missing or matching prefix) and inserted a missing
prefix (which caused the problem of #1431).

This patch now only considers unexpanded IRIs as predefined.

Another minor change ensures that a prefix is always only set,
when it was empty before.
This was not case for the rdfs annotation properties:
label, comment, seeAlso, isDefinedBy