Filter angular brackets in `lookupSerialization`
Same (maybe) problem as in `lookupLogicName`. The IRI passed in as a
parameter is still enclosed in angular brackets which is not expected by
those functions. Filtering the brackets might only address the symptom
of a bigger issue but it fixes the problem at hand and has the least
effects on the remaining codebase.