Add hacky standalone `language` declaration support
As the comments says, theres a proper way to do it by using the
information stored in the `LanguageQual` that is the result of parsing
the `language` declaration. But this would entail expanding a possible
prefix of the `LanguageQual` and I don't have time to implement this
right now. I'll open an issue for this so that I can work on it later.

For now, parsing the DOL examples is more important so I added this hack
to get an intermediate working solution.